Added StringPermutations class and removed StringPermutations console app
