+++ ProblemSolving.sln
@@ -3,11 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.2.32630.192
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "StringPermutations", "StringPermutations\StringPermutations.csproj", "{97276C52-198F-4455-B1C1-A4BBF62D7D87}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebSite", "WebSite\WebSite.csproj", "{CEA0F06C-DE8A-4080-BC10-678151248C47}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebSite", "WebSite\WebSite.csproj", "{CEA0F06C-DE8A-4080-BC10-678151248C47}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Problems", "Problems\Problems.csproj", "{D321C0C0-6096-4CB4-8D06-B68A4274BC51}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Problems", "Problems\Problems.csproj", "{D321C0C0-6096-4CB4-8D06-B68A4274BC51}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,10 +13,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{97276C52-198F-4455-B1C1-A4BBF62D7D87}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{97276C52-198F-4455-B1C1-A4BBF62D7D87}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{97276C52-198F-4455-B1C1-A4BBF62D7D87}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{97276C52-198F-4455-B1C1-A4BBF62D7D87}.Release|Any CPU.Build.0 = Release|Any CPU
 		{CEA0F06C-DE8A-4080-BC10-678151248C47}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{CEA0F06C-DE8A-4080-BC10-678151248C47}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CEA0F06C-DE8A-4080-BC10-678151248C47}.Release|Any CPU.ActiveCfg = Release|Any CPU
